refactor: unify logging imports and init sequence
